chore: move to pyproject and poetry
