--- .coveragerc
@@ -1,5 +1,0 @@
-[run]
-omit =
-    renku/conftest.py
-    renku/data/*
-
