+++ .gitignore
@@ -1,6 +1,3 @@
-# Automatically generated by setuptools_scm
-renku/version.py
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 .pytest_cache/
