--- .tx/config
@@ -1,43 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright 2017-2021- Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
-# Eidgenössische Technische Hochschule Zürich (ETHZ).
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# TODO: Transifex integration
-#
-# 1) Create message catalog:
-#    $ python setup.py extract_messages
-#    $ python setup.py init_catalog -l <lang>
-#    $ python setup.py compile_catalog
-# 2) Ensure project has been created on Transifex under the SDSC
-#    organisation.
-# 3) Install the transifex-client
-#    $ pip install transifex-client
-# 4) Push source (.pot) and translations (.po) to Transifex
-#    $ tx push -s -t
-# 5) Pull translations for a single language from Transifex
-#    $ tx pull -l <lang>
-# 6) Pull translations for all languages from Transifex
-#    $ tx pull -a
-
-[main]
-host = https://www.transifex.com
-
-[sdsc.renku-messages]
-file_filter = renku/translations/<lang>/LC_MESSAGES/messages.po
-source_file = renku/translations/messages.pot
-source_lang = en
-type = PO
