--- MANIFEST.in
@@ -1,75 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright 2017-2021 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
-# Eidgenössische Technische Hochschule Zürich (ETHZ).
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Check manifest will not automatically add these two files:
-include renku/service/.env-example
-include .dockerignore
-include .editorconfig
-include .tx/config
-include *.md
-prune docs/_build
-recursive-include renku *.po *.pot *.mo
-recursive-include renku *.py *.tmpl
-# added by check_manifest.py
-chartpress.yaml
-include .coveragerc
-include *.enc
-include *.py
-include *.rst
-include *.sh
-include *.txt
-include *.yaml
-include *.yml
-include babel.ini
-include Dockerfile*
-include LICENSE
-include Makefile
-include Pipfile Pipfile.lock
-include pyproject.toml
-include pytest.ini
-include renku/data/defaults.ini
-include snap/snapcraft.yaml
-prune .env
-prune .github
-prune design
-recursive-include helm-chart .helmignore *.yaml *.lock *.tpl *.rst *.tgz *.json
-recursive-include .github CODEOWNERS
-recursive-include .travis *.sh
-recursive-include docs *.bat
-recursive-include docs *.css
-recursive-include docs *.html
-recursive-include docs *.ico
-recursive-include docs *.png
-recursive-include docs *.gif
-recursive-include docs *.py
-recursive-include docs *.rst
-recursive-include docs *.svg
-recursive-include docs *.txt
-recursive-include docs Makefile
-recursive-include examples *.py
-recursive-include examples *.sh
-recursive-include tests *.py *.gz *.yml *.json
-recursive-include renku *.default
-recursive-include renku *.html
-recursive-include renku *.json
-recursive-include renku *.sh
-recursive-include renku *.txt
-recursive-include renku *.yml
-recursive-include renku Dockerfile
-recursive-include renku *.yaml
-recursive-include renku/templates *
