--- Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-renku = {editable = true, extras = ["all"], path = "."}
-
-[dev-packages]
-chartpress = "==1.3.0"
-
-[requires]
-python_version = "3.7"
-
-[scripts]
-tests = "./run-tests.sh -t -s -d"
