+++ renku/service/controllers/api/__init__.py
@@ -15,19 +15,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-[pytest]
-addopts = --flake8 --black --doctest-glob="*.rst" --doctest-modules --cov=renku --cov-config .coveragerc --cov-report=term-missing
-doctest_optionflags = ALLOW_UNICODE ALLOW_BYTES DONT_ACCEPT_TRUE_FOR_1 ELLIPSIS IGNORE_EXCEPTION_DETAIL
-flake8-ignore =
-    *.py E121 E126 E203 E226 E231 W503 W504
-    docs/conf.py ALL
-flake8-max-line-length = 120
-testpaths = docs tests renku conftest.py
-markers =
-    integration: mark a test as a integration.
-    service: mark a test as service test.
-    jobs: mark a test as a job test.
-    migration: mark a test as a migration test.
-    shelled: mark a test as a shelled test.
-    serial: mark a test that can not be run in parallel
+"""Renku service controllers api."""
