+++ renku/service/jobs/__init__.py
@@ -15,20 +15,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-# Extraction from Python source files
-
-[python: **.py]
-encoding = utf-8
-
-# Extraction from Jinja2 templates
-
-[jinja2: **/templates/**.html]
-encoding = utf-8
-extensions = jinja2.ext.autoescape, jinja2.ext.with_
-
-# Extraction from JavaScript files
-
-[javascript: **.js]
-encoding = utf-8
-extract_messages = $._, jQuery._
+"""Renku service jobs."""
